+++ common/src/main/java/com/hammy275/immersivemc/client/immersive/ImmersiveBuilder.java
@@ -3,8 +3,8 @@ package com.hammy275.immersivemc.client.immersive;
 import com.hammy275.immersivemc.client.config.ClientConstants;
 import com.hammy275.immersivemc.client.immersive.info.BuiltImmersiveInfo;
 import com.hammy275.immersivemc.common.immersive.CheckerFunction;
-import com.hammy275.immersivemc.common.immersive.HandlerStorage;
-import com.hammy275.immersivemc.common.immersive.ImmersiveHandler;
+import com.hammy275.immersivemc.common.immersive.storage.HandlerStorage;
+import com.hammy275.immersivemc.common.immersive.handler.ImmersiveHandler;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Vec3i;
 import net.minecraft.world.level.Level;
