+++ common/src/main/java/com/hammy275/immersivemc/client/immersive/Immersives.java
@@ -6,10 +6,10 @@ import com.hammy275.immersivemc.client.immersive.info.AnvilData;
 import com.hammy275.immersivemc.client.immersive.info.ChestLikeData;
 import com.hammy275.immersivemc.client.immersive.info.EnchantingData;
 import com.hammy275.immersivemc.common.config.ActiveConfig;
-import com.hammy275.immersivemc.common.immersive.AnvilStorage;
-import com.hammy275.immersivemc.common.immersive.ETableStorage;
+import com.hammy275.immersivemc.common.immersive.storage.AnvilStorage;
+import com.hammy275.immersivemc.common.immersive.storage.ETableStorage;
 import com.hammy275.immersivemc.common.immersive.ImmersiveCheckers;
-import com.hammy275.immersivemc.common.immersive.ImmersiveHandlers;
+import com.hammy275.immersivemc.common.immersive.handler.ImmersiveHandlers;
 import com.hammy275.immersivemc.common.network.Network;
 import com.hammy275.immersivemc.common.network.packet.InteractPacket;
 import com.hammy275.immersivemc.common.network.packet.SwapPacket;
