+++ common/src/main/java/com/hammy275/immersivemc/common/immersive/handler/HandlerUtil.java
@@ -1,5 +1,6 @@
-package com.hammy275.immersivemc.common.immersive;
+package com.hammy275.immersivemc.common.immersive.handler;
 
+import com.hammy275.immersivemc.common.immersive.storage.ListOfItemsStorage;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.Container;
 import net.minecraft.world.item.ItemStack;
