+++ common/src/main/java/com/hammy275/immersivemc/common/network/NetworkClientHandlers.java
@@ -5,8 +5,8 @@ import com.hammy275.immersivemc.client.immersive.AbstractImmersive;
 import com.hammy275.immersivemc.client.immersive.BuiltImmersive;
 import com.hammy275.immersivemc.client.immersive.Immersives;
 import com.hammy275.immersivemc.client.immersive.info.*;
-import com.hammy275.immersivemc.common.immersive.HandlerStorage;
-import com.hammy275.immersivemc.common.immersive.ListOfItemsStorage;
+import com.hammy275.immersivemc.common.immersive.storage.HandlerStorage;
+import com.hammy275.immersivemc.common.immersive.storage.ListOfItemsStorage;
 import com.hammy275.immersivemc.common.network.packet.BeaconDataPacket;
 import net.minecraft.core.BlockPos;
 import net.minecraft.resources.ResourceLocation;
