+++ common/src/main/java/com/hammy275/immersivemc/common/network/packet/FetchInventoryPacket.java
@@ -3,6 +3,10 @@ package com.hammy275.immersivemc.common.network.packet;
 import com.hammy275.immersivemc.ImmersiveMC;
 import com.hammy275.immersivemc.common.compat.Lootr;
 import com.hammy275.immersivemc.common.immersive.*;
+import com.hammy275.immersivemc.common.immersive.handler.ImmersiveHandler;
+import com.hammy275.immersivemc.common.immersive.handler.ImmersiveHandlers;
+import com.hammy275.immersivemc.common.immersive.storage.HandlerStorage;
+import com.hammy275.immersivemc.common.immersive.storage.ListOfItemsStorage;
 import com.hammy275.immersivemc.common.network.Network;
 import com.hammy275.immersivemc.common.network.NetworkClientHandlers;
 import com.hammy275.immersivemc.common.network.NetworkUtil;
