+++ common/src/main/java/com/hammy275/immersivemc/common/network/packet/SwapPacket.java
@@ -4,8 +4,8 @@ import com.hammy275.immersivemc.client.SafeClientUtil;
 import com.hammy275.immersivemc.common.config.ActiveConfig;
 import com.hammy275.immersivemc.common.config.PlacementMode;
 import com.hammy275.immersivemc.common.immersive.ImmersiveCheckers;
-import com.hammy275.immersivemc.common.immersive.ImmersiveHandler;
-import com.hammy275.immersivemc.common.immersive.ImmersiveHandlers;
+import com.hammy275.immersivemc.common.immersive.handler.ImmersiveHandler;
+import com.hammy275.immersivemc.common.immersive.handler.ImmersiveHandlers;
 import com.hammy275.immersivemc.common.network.NetworkUtil;
 import com.hammy275.immersivemc.server.swap.Swap;
 import dev.architectury.networking.NetworkManager;
